Fix latex-export pylint findings (break import cycles + per-file)
The LaTeX exporter (#222) carries pylint debt that `pylint proof_frog` on
main passes only by luck: pylint's package-mode traversal never closes the
latex-internal import cycles, so the cyclic-import (R0401) and the per-file
findings it cascades stay quiet. Add any sibling package to the module graph
(e.g. the easycrypt exporter) and pylint discovers the cycle, surfacing
R0401 plus too-many-arguments/positional and protected-access on the same
modules -- turning `make lint` red.

Fix the root causes rather than suppressing the symptom:

- Break the `exporter <-> proof_renderer` cycle by extracting the leaf
  document-assembly helpers (`assemble`, `_fragment`, `_package_lines`) into
  a new `document.py`; both import it one-directionally. `assemble` was not
  imported by any test, so this is API-safe.
- Break the `expr_renderer <-> type_renderer` cycle by co-locating
  `TypeRenderer` (which composes `ExprRenderer`, which renders a `Type` via
  `TypeRenderer`) in `expr_renderer.py`; `type_renderer.py` re-exports it for
  backward compatibility (module_renderer and test_type_renderer still import
  from there).
- Relocate the two `# pylint: disable=protected-access` comments onto the
  access line (they were on the closing-paren line, so pylint ignored them).
- Add `too-many-arguments,too-many-positional-arguments` disables to the two
  public render entry points (`export_file`, `render_proof`).

`pylint proof_frog` stays 10.00/10 on main and is now also clean when the
easycrypt export package is in the graph. black/mypy clean; 211 latex tests
pass.
